Moved logic to a separate file, added the successful message again
